code improvements

+++ app/Http/Controllers/API/UserController.php
@@ -2,7 +2,13 @@
 
 namespace App\Http\Controllers\API;
 
-use Illuminate\Http\Request;
+use App\Http\Requests\Gate\OwnerAndAdminRequest;
+
+use App\Http\Requests\User\RecepCapRequest;
+use App\Http\Requests\User\CreateRequest;
+use App\Http\Requests\User\UpdateRequest;
+use App\Http\Requests\User\UpdateProfileRequest;
+use App\Http\Requests\Gate\AllRequest;
 use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\Hash;
 use App\Notifications\EmailVerificationForNewRegistered;
@@ -17,35 +23,21 @@ class UserController extends Controller
         $this->middleware(['auth:api', 'verified', 'two_factor_auth'])->except('checkTwoFactExpired');
     }
 
-    public function index() {
-        if(!\Gate::allows('superAdmin') && !\Gate::allows('hotelOwner'))
-          return die('not allowed');
-
-        if(\Gate::allows('hotelOwner')) 
-            return User::whereIn('id', Helpers::recep())->where('role', 'hotel_receptionist')->orderBy('created_at', 'desc')->get();
-        
-        if(\Gate::allows('superAdmin'))
-		    return User::where('role', '!=', 'super_admin')->orderBy('created_at', 'desc')->get();  	
+    public function index(OwnerAndAdminRequest $ownerAndAdminRequest) 
+    {
+        if(\Gate::allows('hotelOwner'))  return User::whereIn('id', Helpers::recep())->where('role', 'hotel_receptionist')->orderBy('created_at', 'desc')->get();   
+        if(\Gate::allows('superAdmin')) return User::where('role', '!=', 'super_admin')->orderBy('created_at', 'desc')->get();  	
     }
 
-    public function create(Request $request) {
-        if(!\Gate::allows('superAdmin') && !\Gate::allows('hotelOwner'))
-          return die('not allowed');
-
+    public function create(OwnerAndAdminRequest $ownerAndAdminRequest, CreateRequest $request) 
+    {
         $user = null;
-        $this->isRecep = (($request['role']=='hotel_receptionist')? true : false );
-        $data = [
-                'fullname' => 'required|string|max:191',
-                'email'    => 'required|string|email|max:191|unique:users',
-                'role'     => 'required|string|max:191',
-                'status'   => 'required|string|max:191'
-                ];
-        
+        $this->isRecep = (($request['role']=='hotel_receptionist')? true : false );        
         $dataCreate = [
-                    'name'     => $request['fullname'],
-                    'email'    => $request['email'],
-                    'role'     => $request['role'],
-                    'status'   => $request['status'],
+                    'name'     => $request->fullname,
+                    'email'    => $request->email,
+                    'role'     => $request->role,
+                    'status'   => $request->status,
                     'password' => Hash::make('pabook@123')
                     ];
         
@@ -53,12 +45,10 @@ class UserController extends Controller
         
         if(\Gate::allows('superAdmin')) {
             if($this->isRecep) {
-                $data['assignTo'] = 'required|numeric|min:1';
-                $userMeta['user_id'] = $request['assignTo'];   
+                $userMeta['user_id'] = $request->assignTo;   
             }   
         }
-
-    	$this->validate($request, $data);
+    	$validated = $request->validated();
 
         if(\Gate::allows('hotelOwner')) {
             $userMeta['user_id'] = auth('api')->user()->id;
@@ -75,64 +65,43 @@ class UserController extends Controller
         return ( ($user!=null)? $user : die('something went wrong!') );
     }
 
-    public function show($id) {
-        if(!\Gate::allows('superAdmin') && !\Gate::allows('hotelOwner'))
-            return die('not allowed');
-
-        if(\Gate::allows('hotelOwner')) 
-            return User::whereIn('id', Helpers::recep())->where('id', $id)->where('role', 'hotel_receptionist')->first();
-        
-        if(\Gate::allows('superAdmin'))
-            return User::where('id', $id)->where('role', '!=', 'super_admin')->first();
+    public function show(OwnerAndAdminRequest $ownerAndAdminRequest, $id) 
+    {
+        if(\Gate::allows('hotelOwner')) return User::whereIn('id', Helpers::recep())->where('id', $id)->where('role', 'hotel_receptionist')->first();   
+        if(\Gate::allows('superAdmin')) return User::where('id', $id)->where('role', '!=', 'super_admin')->first();
     }
 
-    public function update(Request $request, $id) {
-        if(!\Gate::allows('superAdmin') && !\Gate::allows('hotelOwner'))
-            return die('not allowed');
-
+    public function update(OwnerAndAdminRequest $ownerAndAdminRequest, UpdateRequest $request, $id) 
+    {
+        $validated = $request->validated();
         return $this->validateUpdate($request, $id);
     }
 
-    public function destroy($id) {
-        if(!\Gate::allows('superAdmin') && !\Gate::allows('hotelOwner'))
-            return die('not allowed');
-
-        if(\Gate::allows('hotelOwner'))
-            return User::whereIn('id', Helpers::recep())->where('id', $id)->delete();
-
-        if(\Gate::allows('superAdmin')) 
-    	   return User::where('id', $id)->delete();
+    public function destroy(OwnerAndAdminRequest $ownerAndAdminRequest, $id) 
+    {
+        if(\Gate::allows('hotelOwner')) return User::whereIn('id', Helpers::recep())->where('id', $id)->delete();
+        if(\Gate::allows('superAdmin')) return User::where('id', $id)->delete();
     }
 
-    public function profile() {
-        if(!\Gate::allows('superAdmin') && !\Gate::allows('hotelOwner') && !\Gate::allows('hotelReceptionist'))
-            return die('not allowed');
-
+    public function profile(AllRequest $allRequest) 
+    {
         return auth('api')->user();
     }
 
-    public function updateProfile(Request $request) {
-        if(!\Gate::allows('superAdmin') && !\Gate::allows('hotelOwner') && !\Gate::allows('hotelReceptionist')) 
-            return die('not allowed');
-
+    public function updateProfile(AllRequest $allRequest, UpdateProfileRequest $request) 
+    {
+        $validated = $request->validated();
         return $this->validateUpdate($request, auth('api')->user()->id);
     }
 
-    public function hotelOwner() {
-        if(!\Gate::allows('superAdmin') && !\Gate::allows('hotelOwner'))
-            return die('not allowed');
-
-        if(\Gate::allows('superAdmin'))
-            return User::select('id', 'email')->where('role', 'hotel_owner')->get();
-
-        if(\Gate::allows('hotelOwner'))
-            return User::select('id', 'email')->where('id', auth('api')->user()->id)->where('role', 'hotel_owner')->first();
+    public function hotelOwner(OwnerAndAdminRequest $ownerAndAdminRequest) 
+    {
+        if(\Gate::allows('superAdmin')) return User::select('id', 'email')->where('role', 'hotel_owner')->get();
+        if(\Gate::allows('hotelOwner')) return User::select('id', 'email')->where('id', auth('api')->user()->id)->where('role', 'hotel_owner')->first();
     }
 
-    public function hotelReceptionist($id=null) {
-        if(!\Gate::allows('superAdmin') && !\Gate::allows('hotelOwner'))
-            return die('not allowed');
-
+    public function hotelReceptionist(OwnerAndAdminRequest $ownerAndAdminRequest, $id=null) 
+    {
         if(\Gate::allows('superAdmin')) {
             $recep = UserMeta::select('value')->where('meta_key', 'receptionist_id')->where('user_id', $id)->get()->toArray();
             return User::where('role', 'hotel_receptionist')->where('status', 'active')->whereIn('id', $recep)->get();
@@ -141,26 +110,20 @@ class UserController extends Controller
             return User::where('role', 'hotel_receptionist')->where('status', 'active')->whereIn('id', Helpers::recep())->get();
     }
 
-    public function recapCap(Request $request, $action) {
-        if(!\Gate::allows('superAdmin') && !\Gate::allows('hotelOwner'))
-          return die('not allowed');
-
-        $data = [
-                'recep' => 'required|numeric|min:1'
-                ];
-
+    public function recepCap(OwnerAndAdminRequest $ownerAndAdminRequest, RecepCapRequest $request, $action) 
+    {
         $hotel_ids = array();
-        foreach ($request['assignHotel'] as $key => $value) {
+        foreach ($request->assignHotel as $key => $value) {
             array_push($hotel_ids, $value['id']);
         }        
 
         $userMeta = [
-                'user_id'  => $request['recep'],
+                'user_id'  => $request->recep,
                 'meta_key' => 'assign_to_hotel',
                 'value'    => json_encode($hotel_ids)
                 ];        
         
-        $this->validate($request, $data);
+        $validated = $request->validated();
                 
         if($action=='add') {
             $isMetakeyExist = UserMeta::where('user_id', $request['recep'])->where('meta_key', 'assign_to_hotel')->first();
@@ -177,7 +140,8 @@ class UserController extends Controller
     /**
      * Custom query
      */
-    public function checkTwoFactExpired() {
+    public function checkTwoFactExpired() 
+    {
         $user = auth('api')->user();
         if ($user->two_factor_expiry < \Carbon\Carbon::now()) {
             return 'reload';
@@ -188,44 +152,31 @@ class UserController extends Controller
     /**
     *  Validate update for single user page and user profile
     */
-    private function validateUpdate(Request $request, $id) {
-        $data = [
-                'fullname' => 'required|string|max:191',
-                'email'    => 'required|string|email|max:191|unique:users,email,'.$id,
-                'role'     => 'required|string|max:191',
-                'status'   => 'required|string|max:191'
-                ];
+    private function validateUpdate($request, $id) 
+    {
         $dataUpdate = [
-                      'name'     => $request['fullname'],
-                      'email'    => $request['email'],
-                      'role'     => $request['role'],
-                      'status'   => $request['status']
+                      'name'     => $request->fullname,
+                      'email'    => $request->email,
+                      'role'     => $request->role,
+                      'status'   => $request->status
                       ]; 
 
-        $customMessages = [
-                        'match_old_password' => 'The :attribute field doesn\'t match the current password.'
-                        ];              
-
-        if($request['changePass']) {
-            $data['password'] = 'required|string|min:6';
-            $data['old_password'] = 'required|string|min:6|match_old_password:'.$request['old_password'].','.auth('api')->user()->password; 
-            $dataUpdate['password'] = Hash::make($request['password']);
+        if($request->changePass) {
+            $dataUpdate['password'] = Hash::make($request->password);
         }
 
         $emailChange = false;
         if($id==auth('api')->user()->id) {
-            if($request['email']!=auth('api')->user()->email) {
+            if($request->email!=auth('api')->user()->email) {
                 $dataUpdate['email_verified_at'] = null;
                 $emailChange = true;
             }
         }else{
-            if($request['email']!=User::where('id', $id)->first()->email) {
+            if($request->email!=User::where('id', $id)->first()->email) {
                 $dataUpdate['email_verified_at'] = null;
                 $emailChange = true;
             }
         }
-
-        $this->validate($request, $data, $customMessages);
 
         $user = User::where('id', $id)->update($dataUpdate); 
         if($emailChange) {
